tecfood/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cloudinary==1.40.0
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
